Use schedule-sensitive facilities for rationalization
